agent_b/requirements.txt
# === Core Dependencies ===
google-generativeai>=0.5.4
requests>=2.32.3

# === Configuration & Utils ===
python-dotenv>=1.0.1

# === Development & Testing ===
pytest>=8.3.3
pytest-asyncio>=0.23.8
coverage>=7.6.2

# === Optional for enhanced functionality ===
rich>=13.7.1  # CLI colors and formatting
